refactor: reorganize pages into lab-admin and field-tech folders
