+++ src/pages/field-tech/ReportDetails.tsx
